🔐 Added HTML login template with error handling 🔐
This commit introduces an HTML login template that allows users to enter their email and password to log in. If there is an error, such as an invalid email or password, it will be displayed in red on the page.

The template includes a form that uses the POST method to send the login credentials to the '/' route. It also provides a link to the registration page for new users.