📝 Added HTML registration template with error handling 📝
This commit introduces an HTML registration template that allows new users to sign up by providing their email and password. If there is an error, such as an existing user with the same email, it will be displayed in red on the page.

The template includes a form that uses the POST method to send the registration data to the '/register' route. It also provides a link to the login page for users who already have an account.